+++ crates/ruma-identifiers-validation/src/lib.rs
@@ -15,6 +15,7 @@ pub mod room_version_id;
 pub mod server_name;
 pub mod session_id;
 pub mod user_id;
+pub mod voip_version_id;
 
 pub use error::Error;